+++ source/common/common/standard/logger_impl.h
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace Envoy {
+namespace Logger {
+
+#define GENERATE_LOGGER(X) StandardLogger(#X),
+
+} // namespace Logger
+} // namespace Envoy
